+++ src/fund/index.ts
@@ -1,7 +1,7 @@
 export { FundButton } from './components/FundButton';
 export { getCoinbaseSmartWalletFundUrl } from './utils/getCoinbaseSmartWalletFundUrl';
 export { getOnrampBuyUrl } from './utils/getOnrampBuyUrl';
-export { setupEventSubscriptions } from './utils/setupEventSubscriptions';
+export { setupOnrampEventListeners } from './utils/setupOnrampEventListeners';
 
 export type {
   GetOnrampUrlWithProjectIdParams,
